make things TS-y on CORE, shush a key warningabout the project contact, and remove some unused params
